Stack from [ghstack](https://github.com/ezyang/ghstack):
* **#49910 [WIP] Demo for caching serialization**

Demo to demonstrate potential speedup if we cache serialization in RPC.

With this setup where we have several large networks being serialized to be send over network with rpc.remote, we can cache the serialization the first time and this results in some speedup. In particular, total execution time goes from 2s to 1s; each (cached) rpc.remote invocaton goes from ~0.16s -> ~0.05s.

Differential Revision: [D25720408](https://our.internmc.facebook.com/intern/diff/D25720408/)